requirements.txt
certifi==2022.6.15
charset-normalizer==2.1.1
clip==1.0
et-xmlfile==1.1.0
ftfy==6.1.1
idna==3.3
numpy==1.23.2
openai==0.23.0
openpyxl==3.0.10
pandas==1.4.3
pandas-stubs==1.4.3.220822
Pillow==9.2.0
python-dateutil==2.8.2
pytz==2022.2.1
regex==2022.8.17
requests==2.28.1
six==1.16.0
torch==1.12.1
torchvision==0.13.1
tqdm==4.64.0
types-pytz==2022.2.1.0
typing-extensions==4.3.0
urllib3==1.26.12
wcwidth==0.2.5
